+++ src/Telegram.Bot/Exceptions/ApiExceptionParser.cs
@@ -1,5 +1,6 @@
-﻿using System;
+using System;
 using System.Linq;
+using System.Reflection;
 using System.Text.RegularExpressions;
 using Telegram.Bot.Types;
 
@@ -7,19 +8,23 @@ namespace Telegram.Bot.Exceptions
 {
     internal static class ApiExceptionParser
     {
-        private static readonly IApiExceptionInfo<ApiRequestException>[] ExceptionInfos = {
+        private static readonly IApiExceptionInfo<ApiRequestException>[] BadRequestExceptionInfos =
+        {
+            #region BadRequest Exceptions
+
             new BadRequestExceptionInfo<ChatNotFoundException>("chat not found"),
             new BadRequestExceptionInfo<UserNotFoundException>("user not found"),
-            new BadRequestExceptionInfo<InvalidUserIdException>("USER_ID_INVALID"),
-            new BadRequestExceptionInfo<InvalidQueryIdException>("QUERY_ID_INVALID"),
-
-            #region Stickers
-
-            new BadRequestExceptionInfo<InvalidStickerSetNameException>("sticker set name invalid"),
-            new BadRequestExceptionInfo<InvalidStickerEmojisException>("invalid sticker emojis"),
-            new BadRequestExceptionInfo<InvalidStickerDimensionsException>("STICKER_PNG_DIMENSIONS"),
+            new BadRequestExceptionInfo<WrongChatTypeException>("method is available (for supergroup and channel chats only|only for supergroups)"),
+            new BadRequestExceptionInfo<ChatNotModifiedException>("CHAT_NOT_MODIFIED"),
             new BadRequestExceptionInfo<StickerSetNameExistsException>("sticker set name is already occupied"),
             new BadRequestExceptionInfo<StickerSetNotModifiedException>("STICKERSET_NOT_MODIFIED"),
+            new BadRequestExceptionInfo<ContactRequestException>("phone number can be requested in a private chats only"),
+            new BadRequestExceptionInfo<MessageIsNotModifiedException>("message is not modified"),
+            new BadRequestExceptionInfo<ChatDescriptionIsNotModifiedException>("chat description is not modified"),
+            // ToDo: BotRestrictedException test case
+            new BadRequestExceptionInfo<BotRestrictedException>("have no rights to send a message"),
+            // ToDo: NotEnoughRightsException test case
+            new BadRequestExceptionInfo<NotEnoughRightsException>("not enough rights to restrict/unrestrict chat member"),
 
             #endregion
 
@@ -30,54 +35,80 @@ namespace Telegram.Bot.Exceptions
             new BadRequestExceptionInfo<InvalidGameShortNameException>("wrong game short name specified"),
 
             #endregion
-            
-            new BadRequestExceptionInfo<ContactRequestException>("phone number can be requested in a private chats only"),
 
-            new ForbiddenExceptionInfo<ChatNotInitiatedException>("bot can't initiate conversation with a user"),
+            #region InvalidParameter Exceptions
 
+            new BadRequestExceptionInfo<InvalidUserIdException>("USER_ID_INVALID"),
+            new BadRequestExceptionInfo<InvalidQueryIdException>("QUERY_ID_INVALID"),
+            new BadRequestExceptionInfo<InvalidStickerSetNameException>("sticker set name invalid"),
+            new BadRequestExceptionInfo<InvalidStickerEmojisException>("invalid sticker emojis"),
+            new BadRequestExceptionInfo<InvalidStickerDimensionsException>("STICKER_PNG_DIMENSIONS"),
             new BadRequestExceptionInfo<InvalidParameterException>($@"\w{{3,}} Request: invalid (?<{InvalidParameterException.ParamGroupName}>[\w|\s]+)$"),
             new BadRequestExceptionInfo<InvalidParameterException>($@"\w{{3,}} Request: (?<{InvalidParameterException.ParamGroupName}>[\w|\s]+) invalid$"),
+            // ToDo: rename MissingParameterException to EmptyParameterException
+            new BadRequestExceptionInfo<MissingParameterException>($@"\w{{3,}} Request: (?<{InvalidParameterException.ParamGroupName}>[\w|\s]+) is empty"),
 
-            new BadRequestExceptionInfo<MessageIsNotModifiedException>("message is not modified"),
+            #endregion
+        };
+
+        private static readonly IApiExceptionInfo<ApiRequestException>[] ForibddenExceptionInfos =
+        {
+            #region Forbidden Exceptions
+            new ForbiddenExceptionInfo<ChatNotInitiatedException>("bot can't initiate conversation with a user"),
+            new ForbiddenExceptionInfo<SendMessageToBotException>("bot can't send messages to bots"),
+            new ForbiddenExceptionInfo<BotIsNotMemberException>("bot is not a member of the (supergroup|channel) chat"),
+            // ToDo: BotBlockedException test case
+            new ForbiddenExceptionInfo<BotBlockedException>("bot was blocked by the user"),
+            #endregion
         };
 
         public static ApiRequestException Parse<T>(ApiResponse<T> apiResponse)
         {
-            ApiRequestException exception;
+            var errorMessage = string.Empty;
+            IApiExceptionInfo<ApiRequestException> typeInfo = default;
 
-            var typeInfo = ExceptionInfos
-                .FirstOrDefault(info => Regex.IsMatch(apiResponse.Description, info.ErrorMessageRegex));
-
-            if (typeInfo is null)
+            switch (apiResponse.ErrorCode)
             {
-                exception = new ApiRequestException(apiResponse.Description, apiResponse.ErrorCode, apiResponse.Parameters);
-            }
-            else
-            {
-                string errorMessage;
-                bool isBadRequestError = typeInfo.ErrorCode == BadRequestException.BadRequestErrorCode;
-
-                if (isBadRequestError)
-                {
+                case BadRequestException.BadRequestErrorCode:
                     errorMessage = TruncateBadRequestErrorDescription(apiResponse.Description);
 
-                    if (typeInfo.Type == typeof(InvalidParameterException))
+                    typeInfo =
+                        BadRequestExceptionInfos.FirstOrDefault(info => Regex.IsMatch(apiResponse.Description, info.ErrorMessageRegex))
+                        ?? new BadRequestExceptionInfo<BadRequestException>(errorMessage);
+
+                    if (IsAssignableFrom<InvalidParameterException>(typeInfo.Type))
                     {
                         string paramName = Regex.Match(apiResponse.Description, typeInfo.ErrorMessageRegex)
                             .Groups[InvalidParameterException.ParamGroupName]
                             .Value;
-                        exception = new InvalidParameterException(paramName, errorMessage);
+
+                        return string.IsNullOrEmpty(paramName)
+                            ? Activator.CreateInstance(typeInfo.Type, errorMessage) as ApiRequestException
+                            : Activator.CreateInstance(typeInfo.Type, paramName, errorMessage) as ApiRequestException;
                     }
-                    else
-                        exception = Activator.CreateInstance(typeInfo.Type, errorMessage) as ApiRequestException;
-                }
-                else
-                {
+
+                    return Activator.CreateInstance(typeInfo.Type, errorMessage) as ApiRequestException;
+
+                case ForbiddenException.ForbiddenErrorCode:
                     errorMessage = TruncateForbiddenErrorDescription(apiResponse.Description);
-                    exception = Activator.CreateInstance(typeInfo.Type, errorMessage) as ApiRequestException;
-                }
+
+                    typeInfo =
+                        ForibddenExceptionInfos.FirstOrDefault(info => Regex.IsMatch(apiResponse.Description, info.ErrorMessageRegex))
+                        ?? new ForbiddenExceptionInfo<ForbiddenException>(errorMessage);
+
+                    return Activator.CreateInstance(typeInfo.Type, errorMessage) as ApiRequestException;
+
+                case TooManyRequestsException.TooManyRequestsErrorCode:
+                    errorMessage = TruncateTooManyRequestsErrorDescription(apiResponse.Description);
+
+                    return new TooManyRequestsException(errorMessage, apiResponse.Parameters);
+
+                default:
+                    return new ApiRequestException(apiResponse.Description, apiResponse.ErrorCode, apiResponse.Parameters);
             }
-            return exception;
+
+            bool IsAssignableFrom<parentT>(Type derivedT) =>
+                typeof(parentT).GetTypeInfo().IsAssignableFrom(derivedT.GetTypeInfo());
         }
 
         private static string TruncateBadRequestErrorDescription(string message) =>
@@ -85,6 +116,9 @@ namespace Telegram.Bot.Exceptions
 
         private static string TruncateForbiddenErrorDescription(string message) =>
             TryTruncateErrorDescription(message, ForbiddenException.ForbiddenErrorDescription);
+
+        private static string TruncateTooManyRequestsErrorDescription(string message) =>
+            TryTruncateErrorDescription(message, TooManyRequestsException.TooManyRequestsErrorDescription);
 
         private static string TryTruncateErrorDescription(string message, string description)
         {