+++ test/Telegram.Bot.Tests.Integ/Exceptions/ApiExceptionsTests.cs
@@ -1,26 +1,31 @@
-﻿using System.Linq;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Telegram.Bot.Exceptions;
 using Telegram.Bot.Tests.Integ.Framework;
+using Telegram.Bot.Tests.Integ.Framework.Fixtures;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 using Telegram.Bot.Types.InlineQueryResults;
 using Telegram.Bot.Types.ReplyMarkups;
 using Xunit;
+using static Telegram.Bot.Tests.Integ.Exceptions.ApiExceptionsTests;
 
 namespace Telegram.Bot.Tests.Integ.Exceptions
 {
     [Collection(Constants.TestCollections.Exceptions)]
-    [TestCaseOrderer(Constants.TestCaseOrderer, Constants.AssemblyName)]
-    public class ApiExceptionsTests
+    [TestCaseOrderer(Constants.TestCaseOrderer2, Constants.AssemblyName)]
+    public class ApiExceptionsTests : IClassFixture<ExceptionsFixture>
     {
         private ITelegramBotClient BotClient => _fixture.BotClient;
 
         private readonly TestsFixture _fixture;
+        private readonly ExceptionsFixture _classFixture;
 
-        public ApiExceptionsTests(TestsFixture fixture)
+        public ApiExceptionsTests(TestsFixture fixture, ExceptionsFixture classFixture)
         {
             _fixture = fixture;
+            _classFixture = classFixture;
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldThrowChatNotFoundException)]
@@ -29,11 +34,12 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowChatNotFoundException);
 
-            BadRequestException e = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
+            BadRequestException exception = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
                 BotClient.SendTextMessageAsync(0, "test")
             );
 
-            Assert.IsType<ChatNotFoundException>(e);
+            Assert.IsType<ChatNotFoundException>(exception);
+            Assert.Equal("chat not found", exception.Message);
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldThrowInvalidUserIdException)]
@@ -42,18 +48,18 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowInvalidUserIdException);
 
-            BadRequestException e = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
+            BadRequestException exception = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
                 BotClient.PromoteChatMemberAsync(_fixture.SupergroupChat.Id, 123456)
             );
 
-            Assert.IsType<InvalidUserIdException>(e);
+            Assert.IsType<InvalidUserIdException>(exception);
+            Assert.Equal("USER_ID_INVALID", exception.Message);
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldThrowExceptionChatNotInitiatedException)]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
         public async Task Should_Throw_Exception_ChatNotInitiatedException()
         {
-            //ToDo add exception. forward message from another bot. Forbidden: bot can't send messages to bots
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionChatNotInitiatedException,
                 "Forward a message to this chat from a user that never started a chat with this bot");
 
@@ -62,14 +68,35 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
             )).Single();
             await _fixture.UpdateReceiver.DiscardNewUpdatesAsync();
 
-            ForbiddenException e = await Assert.ThrowsAnyAsync<ForbiddenException>(() =>
-                BotClient.SendTextMessageAsync(
-                    forwardedMessageUpdate.Message.ForwardFrom.Id,
-                    $"Error! If you see this message, talk to @{forwardedMessageUpdate.Message.From.Username}"
-                )
+            ForbiddenException exception = await Assert.ThrowsAnyAsync<ForbiddenException>(() =>
+                BotClient.SendTextMessageAsync(forwardedMessageUpdate.Message.ForwardFrom.Id,
+                    $"Error! If you see this message, talk to @{forwardedMessageUpdate.Message.From.Username}")
             );
 
-            Assert.IsType<ChatNotInitiatedException>(e);
+            Assert.IsType<ChatNotInitiatedException>(exception);
+            Assert.Equal("bot can't initiate conversation with a user", exception.Message);
+        }
+
+        [Fact(DisplayName = FactTitles.ShouldThrowExceptionSendMessageToBotException)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
+        [ExecutionOrder(3.1)]
+        public async Task Should_Throw_Exception_SendMessageToBotException()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionSendMessageToBotException,
+                "Forward a message to this chat from a bot");
+
+            Update forwardedMessageUpdate = (await _fixture.UpdateReceiver.GetUpdatesAsync(u =>
+                    u.Message.IsForwarded, updateTypes: UpdateType.Message
+            )).Single();
+            await _fixture.UpdateReceiver.DiscardNewUpdatesAsync();
+
+            ForbiddenException exception = await Assert.ThrowsAnyAsync<ForbiddenException>(() =>
+                BotClient.SendTextMessageAsync(forwardedMessageUpdate.Message.ForwardFrom.Id,
+                    $"Error! If you see this message, talk to @{forwardedMessageUpdate.Message.From.Username}")
+            );
+
+            Assert.IsType<SendMessageToBotException>(exception);
+            Assert.Equal("bot can't send messages to bots", exception.Message);
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldThrowExceptionContactRequestException)]
@@ -83,18 +110,17 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
                 KeyboardButton.WithRequestContact("Share Contact"),
             });
 
-            BadRequestException e = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
-                BotClient.SendTextMessageAsync(
-                    _fixture.SupergroupChat.Id,
-                    "You should never see this message",
+            BadRequestException exception = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
+                BotClient.SendTextMessageAsync(_fixture.SupergroupChat.Id, "You should never see this message",
                     replyMarkup: replyMarkup)
             );
 
-            Assert.IsType<ContactRequestException>(e);
+            Assert.IsType<ContactRequestException>(exception);
+            Assert.Equal("phone number can be requested in a private chats only", exception.Message);
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldThrowExceptionMessageIsNotModifiedException)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.EditMessageText)]
         public async Task Should_Throw_Exception_MessageIsNotModifiedException()
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionMessageIsNotModifiedException);
@@ -104,7 +130,7 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
                 _fixture.SupergroupChat.Id,
                 messageTextToModify);
 
-            BadRequestException e = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
+            BadRequestException exception = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
                 BotClient.EditMessageTextAsync(
                     _fixture.SupergroupChat.Id,
                     message.MessageId,
@@ -112,7 +138,8 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
                 )
             );
 
-            Assert.IsType<MessageIsNotModifiedException>(e);
+            Assert.IsType<MessageIsNotModifiedException>(exception);
+            Assert.Equal("message is not modified", exception.Message);
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldThrowExceptionInvalidQueryIdException)]
@@ -120,6 +147,7 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
         public async Task Should_Throw_Exception_QueryIdInvalidException()
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionInvalidQueryIdException,
+                "Pass anything to bot to invoke inline request",
                 startInlineQuery: true);
 
             Update queryUpdate = await _fixture.UpdateReceiver.GetInlineQueryUpdateAsync();
@@ -137,15 +165,138 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
 
             await Task.Delay(10_000);
 
-            InvalidQueryIdException e = await Assert.ThrowsAnyAsync<InvalidQueryIdException>(() =>
+            InvalidParameterException exception = await Assert.ThrowsAnyAsync<InvalidParameterException>(() =>
                 BotClient.AnswerInlineQueryAsync(
                     inlineQueryId: queryUpdate.InlineQuery.Id,
                     results: results,
-                    cacheTime: 0
-                )
+                    cacheTime: 0)
             );
 
-            Assert.Equal("inline_query_id", e.Parameter);
+            Assert.IsType<InvalidQueryIdException>(exception);
+            Assert.Equal("inline_query_id", exception.Parameter);
+        }
+
+        [Fact(DisplayName = FactTitles.ShouldThrowExceptionMissingParameterException)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
+        [ExecutionOrder(7)]
+        public async Task Should_Throw_Exception_MissingParameterException()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionMissingParameterException);
+
+            InvalidParameterException exception = await Assert.ThrowsAnyAsync<InvalidParameterException>(() =>
+                BotClient.SendTextMessageAsync(
+                    _fixture.SupergroupChat.Id,
+                    string.Empty)
+            );
+
+            Assert.IsType<MissingParameterException>(exception);
+            Assert.Equal("message text", exception.Parameter);
+        }
+
+        [Fact(DisplayName = FactTitles.ShouldThrowExceptionWrongChatTypeException_OutsideSupergroups)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SetChatStickerSet)]
+        [ExecutionOrder(8)]
+        public async Task Should_Throw_Exception_WrongChatTypeException_OutsideSupergroups()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionWrongChatTypeException_OutsideSupergroups);
+
+            const string setName = "EvilMinds";
+
+            BadRequestException exception = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
+                _fixture.BotClient.SetChatStickerSetAsync(_classFixture.PrivateChat.Id, setName)
+            );
+
+            Assert.IsType<WrongChatTypeException>(exception);
+            Assert.Equal("method is available only for supergroups", exception.Message);
+        }
+
+        [Fact(DisplayName = FactTitles.ShouldThrowExceptionWrongChatTypeException_OutsideSupergroupsOrChannels)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.UnbanChatMember)]
+        [ExecutionOrder(9)]
+        public async Task Should_Throw_Exception_WrongChatTypeException_OutsideSupergroupsOrChannels()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionWrongChatTypeException_OutsideSupergroupsOrChannels);
+
+            BadRequestException exception = await Assert.ThrowsAnyAsync<BadRequestException>(() =>
+                _fixture.BotClient.UnbanChatMemberAsync(
+                    _classFixture.PrivateChat.Id,
+                    (int)_classFixture.PrivateChat.Id)
+            );
+
+            Assert.IsType<WrongChatTypeException>(exception);
+            Assert.Equal("method is available for supergroup and channel chats only", exception.Message);
+        }
+
+        [Fact(DisplayName = FactTitles.ShouldThrowExceptionBotBlockedException,
+              Skip = "Requires too much user interaction. Could be implemented with Telegram Client Library.")]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
+        [ExecutionOrder(10)]
+        public async Task Should_Throw_Exception_BotBlockedException()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionBotBlockedException);
+
+            const string message = "This message should not be delivered";
+
+            ApiRequestException exception = await Assert.ThrowsAnyAsync<ApiRequestException>(() =>
+                _fixture.BotClient.SendTextMessageAsync(
+                    _classFixture.PrivateChat.Id,
+                    message)
+            );
+
+            Assert.IsType<BotBlockedException>(exception);
+            Assert.Equal("bot was blocked by the user", exception.Message);
+        }
+
+        [Fact(DisplayName = FactTitles.ShouldThrowExceptionTooManyRequestsException)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
+        [ExecutionOrder(11)]
+        public async Task Should_Throw_Exception_TooManyRequestsException()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionTooManyRequestsException);
+
+            const string message = "To the limit!";
+
+            ApiRequestException exception = await Assert.ThrowsAnyAsync<ApiRequestException>(() =>
+            {
+                IEnumerable<Task<Message>> tasks = Enumerable.Range(0, 100)
+                    .Select(_ =>
+                        _fixture.BotClient.SendTextMessageAsync(
+                            _fixture.SupergroupChat.Id,
+                            message));
+                return Task.WhenAll(tasks);
+            });
+
+            Assert.IsType<TooManyRequestsException>(exception);
+            Assert.StartsWith("retry after", exception.Message);
+
+            await Task.Delay(exception.Parameters.RetryAfter * 1_000);
+        }
+
+        [Fact(DisplayName = FactTitles.ShouldThrowExceptionBotIsNotMemberException)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
+        [ExecutionOrder(12)]
+        public async Task Should_Throw_Exception_BotIsNotMemberException()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowExceptionBotIsNotMemberException);
+
+            string[] groups =
+            {
+                "@publictestgroup",
+                "@telegram"
+            };
+            const string message = "To the limit!";
+
+            foreach (string group in groups)
+            {
+                ApiRequestException exception = await Assert.ThrowsAnyAsync<ApiRequestException>(() =>
+                    _fixture.BotClient.SendTextMessageAsync(
+                        group,
+                        message)
+                );
+
+                Assert.IsType<BotIsNotMemberException>(exception);
+                Assert.Matches("bot is not a member of the (supergroup|channel) chat", exception.Message);
+            }
         }
 
         private static class FactTitles
@@ -158,7 +309,10 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
 
             public const string ShouldThrowExceptionChatNotInitiatedException =
                 "Should throw ChatNotInitiatedException while trying to send message to a user who hasn't " +
-                "started a chat with bot but bot knows about him/her.";
+                "started a chat with bot but bot knows about him/her";
+
+            public const string ShouldThrowExceptionSendMessageToBotException =
+                "Should throw SendMessageToBotException when sending message to another bot";
 
             public const string ShouldThrowExceptionContactRequestException =
                 "Should throw ContactRequestException while asking for user's phone number in non-private " +
@@ -169,8 +323,35 @@ namespace Telegram.Bot.Tests.Integ.Exceptions
                 "with the same text";
 
             public const string ShouldThrowExceptionInvalidQueryIdException =
-                "Should throw InvalidQueryIdException when AnswerInlineQueryAsync called with" +
-                " 10 second delay";
+                "Should throw InvalidQueryIdException when AnswerInlineQueryAsync called with " +
+                "10 second delay";
+
+            public const string ShouldThrowExceptionMissingParameterException =
+                "Should throw MissingParameterException when message text is empty";
+
+            public const string ShouldThrowExceptionWrongChatTypeException_OutsideSupergroups =
+                "Should throw WrongChatTypeException when method is not called for supergroup";
+
+            public const string ShouldThrowExceptionWrongChatTypeException_OutsideSupergroupsOrChannels =
+                "Should throw WrongChatTypeException when method is not called for supergroup or channel";
+
+            public const string ShouldThrowExceptionBotBlockedException =
+                "Should throw BotBlockedException when send message to user, who has blocked bot";
+
+            public const string ShouldThrowExceptionTooManyRequestsException =
+                "Should throw TooManyRequestsException when bot reaches message send limit";
+
+            public const string ShouldThrowExceptionBotIsNotMemberException =
+               "Should throw BotIsNotMemberException when bot sends message to the " +
+                "supergroup or channel in which it was not added";
+        }
+
+        public class ExceptionsFixture : PrivateChatFixture
+        {
+            public ExceptionsFixture(TestsFixture testsFixture)
+                : base(testsFixture, Constants.TestCollections.Exceptions)
+            {
+            }
         }
     }
 }