+++ CHANGELOG.md
@@ -20,9 +20,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Implicit cast of `InlineKeyboardButton[]` to `InlineKeyboardMarkup`
 - Exception `InvalidGameShortNameException`
 - Exception `InvalidQueryIdException`
+- Exception `ChatDescriptionIsNotModifiedException`
+- Exception `ChatNotModifiedException`
+- Exception `SendMessageToBotException`
+- Exception `TooManyRequestsException`
+- Exception `BotIsNotMemberException`
 
 ### Changed
 
+- `BadRequestException` is thrown on `Error 400: Bad request` if more specific exception is not defined
+- `ForbiddenException` is thrown on `Error 403: Forbidden` if more specific exception is not defined
 - Made `callbackGame` parameter of `InlineKeyboardButton.WithCallBackGame` optional
 - `Newtonsoft.Json` updated to version `11.0.2`